csv export: unit tests, documentation, linting
* Separate utils into individual file so that they can be tested without
  importing ckan/ckanapi
